third_party/WebKit/LayoutTests/plugins/npruntime/plugin-scriptable-object-invoke-default-expected.txt

 
This page tests invoking as functions JavaScript objects that are provided by plugins.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL typeof(pluginWithInvokeDefault) should be function. Was object.
PASS pluginWithInvokeDefault() is 1
PASS typeof(pluginWithoutInvokeDefault) is 'object'
FAIL pluginWithoutInvokeDefault() should throw an exception. Was undefined.
